EditPost.js code refactored
