+++ posts/editPosts.js
@@ -1,13 +1,18 @@
 import headerElement from "../header/header.js"
-headerElement()
-const editForm = document.getElementById('edit-post-form')
+import {editPost} from "./postsController.js"
+import editedPost from "./postsView.js"
+
+init()
+
+function init() {
+  headerElement()
+let editForm = document.getElementById('edit-post-form')
 let queryParams = document.location.search;
 let urlParams = new URLSearchParams(queryParams);
 let postTitle = urlParams.get('post_title')
 let userName = urlParams.get('user_name')
 let postBody = urlParams.get('post_body')
 let postId = urlParams.get('post_id')
-
 let postEditTitleInput = document.getElementById('post-title')
 let postEditBodyInput = document.getElementById('post-content')
 let postEditAuthorInput = document.getElementById('post-author')
@@ -16,47 +21,19 @@ postEditTitleInput.value = postTitle
 postEditBodyInput.value = postBody
 postEditAuthorInput.value = userName
 
-
-editForm.addEventListener('submit', (e) => {
+editForm.addEventListener('submit', async (e) => {
     e.preventDefault()
    let authorInputEl =  e.target.author.value
-    fetch(`https://jsonplaceholder.typicode.com/posts/${postId}`, {
-     method: 'PUT',
-     body: JSON.stringify({
-    id: `${postId}`,
+   let newPost = {
     title: `${postEditTitleInput.value}`,
     body: `${postEditBodyInput.value}`,
-    author: `${userName}`,
-    userId: 1,
-  }),
-  headers: {
-    'Content-type': 'application/json; charset=UTF-8',
-  },
-})
-  .then((response) => response.json())
-  .then((json) => {
-    
-    let editedPostDiv = document.createElement('ul')
-    editedPostDiv.classList.add('edited-post')
-
-    let postTitleElement = document.createElement('li')
-    postTitleElement.innerHTML = `<strong> Post name:</strong> ${json.title}`
-
-    let postBodyElement = document.createElement('li')
-    postBodyElement.innerHTML = `<strong>Post content:</strong> ${json.body}`
-
-    let postAuthorElement = document.createElement('li')
-    postAuthorElement.innerHTML = `<strong>Post author:</strong> ${authorInputEl}`
-    let postIdElement = document.createElement('li')
-    postIdElement.innerHTML = `<strong>Post id:</strong> ${postId}`
-    
-    editedPostDiv.append(postTitleElement,postBodyElement,postAuthorElement, postIdElement)
-    document.body.append(editedPostDiv)
+    author:`${authorInputEl}`,
+    id: `${postId}`
+  };
+  let postEdited = await editPost(newPost, postId)
+  
+  editedPost(postEdited)
+    e.target.reset()
 });
-e.target.reset()
-    
 
-})
-
-
-
+}
